+++ applications/ParticleMechanicsApplication/custom_conditions/grid_based_conditions/mpm_grid_line_load_condition_2d.cpp
@@ -17,7 +17,7 @@
 
 // Project includes
 #include "includes/define.h"
-#include "custom_conditions/mpm_line_load_condition_2d.h"
+#include "custom_conditions/grid_based_conditions/mpm_grid_line_load_condition_2d.h"
 #include "utilities/math_utils.h"
 #include "utilities/integration_utilities.h"
