+++ applications/ParticleMechanicsApplication/custom_conditions/grid_based_conditions/mpm_grid_line_load_condition_2d.h
@@ -11,8 +11,9 @@
 //
 
 
-#if !defined(KRATOS_MPM_SURFACE_LOAD_CONDITION_3D_H_INCLUDED )
-#define  KRATOS_MPM_SURFACE_LOAD_CONDITION_3D_H_INCLUDED
+// System includes
+#if !defined(KRATOS_MPM_GRID_LINE_LOAD_CONDITION_2D_H_INCLUDED )
+#define      KRATOS_MPM_GRID_LINE_LOAD_CONDITION_2D_H_INCLUDED
 
 // System includes
 
@@ -21,6 +22,7 @@
 // Project includes
 #include "includes/define.h"
 #include "custom_conditions/mpm_base_load_condition.h"
+#include "includes/variables.h"
 
 namespace Kratos
 {
@@ -44,39 +46,30 @@ namespace Kratos
 ///@name Kratos Classes
 ///@{
 
-class MPMSurfaceLoadCondition3D
+/// Short class definition.
+/** Detail class definition.
+*/
+
+class MPMLineLoadCondition2D
     : public MPMBaseLoadCondition
 {
 public:
-
     ///@name Type Definitions
     ///@{
 
-    // Counted pointer of MPMSurfaceLoadCondition3D
-    KRATOS_CLASS_POINTER_DEFINITION( MPMSurfaceLoadCondition3D );
+    /// Counted pointer of MPMLineLoadCondition2D
+    KRATOS_CLASS_POINTER_DEFINITION( MPMLineLoadCondition2D );
 
     ///@}
     ///@name Life Cycle
     ///@{
 
-    // Constructor void
-    MPMSurfaceLoadCondition3D();
+    /// Default constructor.
+    MPMLineLoadCondition2D( IndexType NewId, GeometryType::Pointer pGeometry );
+    MPMLineLoadCondition2D( IndexType NewId, GeometryType::Pointer pGeometry,  PropertiesType::Pointer pProperties );
 
-    // Constructor using an array of nodes
-    MPMSurfaceLoadCondition3D(
-        IndexType NewId,
-        GeometryType::Pointer pGeometry
-        );
-
-    // Constructor using an array of nodes with properties
-    MPMSurfaceLoadCondition3D(
-        IndexType NewId,
-        GeometryType::Pointer pGeometry,
-        PropertiesType::Pointer pProperties
-        );
-
-    // Destructor
-    ~MPMSurfaceLoadCondition3D() override;
+    /// Destructor.
+    ~MPMLineLoadCondition2D() override;
 
     ///@}
     ///@name Operators
@@ -87,7 +80,6 @@ public:
     ///@name Operations
     ///@{
 
-    // Name Operations
     Condition::Pointer Create(
         IndexType NewId,
         GeometryType::Pointer pGeom,
@@ -114,22 +106,37 @@ public:
     ///@name Input and output
     ///@{
 
+    /// Turn back information as a string.
+//      virtual String Info() const;
+
+    /// Print information about this object.
+//      virtual void PrintInfo(std::ostream& rOStream) const;
+
+    /// Print object's data.
+//      virtual void PrintData(std::ostream& rOStream) const;
+
+
     ///@}
     ///@name Friends
     ///@{
 
-protected:
 
+    ///@}
+
+protected:
     ///@name Protected static Member Variables
     ///@{
+
 
     ///@}
     ///@name Protected member Variables
     ///@{
 
+
     ///@}
     ///@name Protected Operators
     ///@{
+
 
     ///@}
     ///@name Protected Operations
@@ -147,51 +154,51 @@ protected:
         MatrixType& rLeftHandSideMatrix,
         VectorType& rRightHandSideVector,
         ProcessInfo& rCurrentProcessInfo,
-        const bool CalculateStiffnessMatrixFlag,
-        const bool CalculateResidualVectorFlag
+        bool CalculateStiffnessMatrixFlag,
+        bool CalculateResidualVectorFlag
         ) override;
 
     void CalculateAndSubKp(
         Matrix& K,
-        const array_1d<double, 3>& ge,
-        const array_1d<double, 3>& gn,
         const Matrix& DN_De,
         const Vector& N,
         const double Pressure,
-        const double Weight );
-
-    void MakeCrossMatrix(
-        BoundedMatrix<double, 3, 3>& M,
-        const array_1d<double, 3>& U
+        const double IntegrationWeight
         );
 
     void CalculateAndAddPressureForce(
-        VectorType& rResidualVector,
+        VectorType& rRightHandSideVector,
         const Vector& N,
-        const array_1d<double, 3 >& Normal,
+        const array_1d<double, 3>& Normal,
         const double Pressure,
-        const double Weight,
-        const ProcessInfo& rCurrentProcessInfo
+        const double IntegrationWeight
         );
 
     ///@}
     ///@name Protected  Access
     ///@{
 
+
     ///@}
     ///@name Protected Inquiry
     ///@{
+
 
     ///@}
     ///@name Protected LifeCycle
     ///@{
 
+    // A protected default constructor necessary for serialization
+    MPMLineLoadCondition2D() {};
+
+    ///@}
+
 private:
-    ///@name Private static Member Variables
+    ///@name Static Member Variables
     ///@{
 
     ///@}
-    ///@name Private member Variables
+    ///@name Member Variables
     ///@{
 
     ///@}
@@ -202,20 +209,14 @@ private:
     ///@name Private Operations
     ///@{
 
+
     ///@}
     ///@name Private  Access
     ///@{
 
+
     ///@}
     ///@name Private Inquiry
-    ///@{
-
-    ///@}
-    ///@name Private LifeCycle
-    ///@{
-
-    ///@}
-    ///@name Unaccessible methods
     ///@{
 
     ///@}
@@ -235,15 +236,49 @@ private:
     }
 
 
-}; // class MPMSurfaceLoadCondition3D.
+    ///@}
+    ///@name Un accessible methods
+    ///@{
 
+    /// Assignment operator.
+    //MPMLineLoadCondition2D& operator=(const MPMLineLoadCondition2D& rOther);
+
+    /// Copy constructor.
+    //MPMLineLoadCondition2D(const MPMLineLoadCondition2D& rOther);
+
+
+    ///@}
+
+}; // Class MPMLineLoadCondition2D
+
+///@}
 ///@name Type Definitions
 ///@{
+
 
 ///@}
 ///@name Input and output
 ///@{
 
-} // namespace Kratos.
 
-#endif // KRATOS_MPM_SURFACE_LOAD_CONDITION_3D_H_INCLUDED  defined
+/// input stream function
+/*  inline std::istream& operator >> (std::istream& rIStream,
+        MPMLineLoadCondition2D& rThis);
+*/
+/// output stream function
+/*  inline std::ostream& operator << (std::ostream& rOStream,
+        const MPMLineLoadCondition2D& rThis)
+    {
+      rThis.PrintInfo(rOStream);
+      rOStream << std::endl;
+      rThis.PrintData(rOStream);
+
+      return rOStream;
+    }*/
+///@}
+
+}  // namespace Kratos.
+
+#endif // KRATOS_MPM_GRID_LINE_LOAD_CONDITION_2D_H_INCLUDED  defined
+
+
