+++ applications/ParticleMechanicsApplication/custom_conditions/grid_based_conditions/mpm_grid_point_load_condition.cpp
@@ -17,7 +17,7 @@
 
 // Project includes
 #include "includes/define.h"
-#include "custom_conditions/mpm_point_load_condition.h"
+#include "custom_conditions/grid_based_conditions/mpm_grid_point_load_condition.h"
 #include "utilities/math_utils.h"
 #include "utilities/integration_utilities.h"
