+++ applications/ParticleMechanicsApplication/custom_conditions/grid_based_conditions/mpm_grid_point_load_condition.h
@@ -12,8 +12,8 @@
 
 
 // System includes
-#if !defined(KRATOS_MPM_POINT_LOAD_CONDITION_H_INCLUDED )
-#define  KRATOS_MPM_POINT_LOAD_CONDITION_H_INCLUDED
+#if !defined(KRATOS_MPM_GRID_POINT_LOAD_CONDITION_H_INCLUDED )
+#define      KRATOS_MPM_GRID_POINT_LOAD_CONDITION_H_INCLUDED
 
 // System includes
 
@@ -65,15 +65,15 @@ public:
     ///@{
 
     /// Default constructor.
-    MPMPointLoadCondition( 
-        IndexType NewId, 
-        GeometryType::Pointer pGeometry 
+    MPMPointLoadCondition(
+        IndexType NewId,
+        GeometryType::Pointer pGeometry
         );
-    
-    MPMPointLoadCondition( 
-        IndexType NewId, 
-        GeometryType::Pointer pGeometry,  
-        PropertiesType::Pointer pProperties 
+
+    MPMPointLoadCondition(
+        IndexType NewId,
+        GeometryType::Pointer pGeometry,
+        PropertiesType::Pointer pProperties
         );
 
     /// Destructor.
@@ -87,17 +87,17 @@ public:
     ///@}
     ///@name Operations
     ///@{
-    
+
     Condition::Pointer Create(
         IndexType NewId,
         GeometryType::Pointer pGeom,
         PropertiesType::Pointer pProperties
         ) const override;
-    
-    Condition::Pointer Create( 
-        IndexType NewId, 
-        NodesArrayType const& ThisNodes,  
-        PropertiesType::Pointer pProperties 
+
+    Condition::Pointer Create(
+        IndexType NewId,
+        NodesArrayType const& ThisNodes,
+        PropertiesType::Pointer pProperties
         ) const override;
 
     ///@}
@@ -158,23 +158,22 @@ protected:
      * @param CalculateStiffnessMatrixFlag: The flag to set if compute the LHS
      * @param CalculateResidualVectorFlag: The flag to set if compute the RHS
      */
-    void CalculateAll( 
-        MatrixType& rLeftHandSideMatrix, 
+    void CalculateAll(
+        MatrixType& rLeftHandSideMatrix,
         VectorType& rRightHandSideVector,
         ProcessInfo& rCurrentProcessInfo,
         bool CalculateStiffnessMatrixFlag,
-        bool CalculateResidualVectorFlag 
+        bool CalculateResidualVectorFlag
         ) override;
-        
+
     /**
-     * It calcules the integration load for the point load 
+     * It calcules the integration load for the point load
      */
     virtual double GetPointLoadIntegrationWeight();
-        
+
     ///@}
     ///@name Protected  Access
     ///@{
-
 
     ///@}
     ///@name Protected Inquiry
@@ -184,7 +183,7 @@ protected:
     ///@}
     ///@name Protected LifeCycle
     ///@{
-    
+
     // A protected default constructor necessary for serialization
     MPMPointLoadCondition() {};
 
@@ -194,12 +193,9 @@ private:
     ///@name Static Member Variables
     ///@{
 
-
     ///@}
     ///@name Member Variables
     ///@{
-
-
 
     ///@}
     ///@name Private Operators
@@ -235,18 +231,6 @@ private:
         KRATOS_SERIALIZE_LOAD_BASE_CLASS( rSerializer, MPMBaseLoadCondition );
     }
 
-    ///@}
-    ///@name Un accessible methods
-    ///@{
-
-    /// Assignment operator.
-    //MPMPointLoadCondition& operator=(const MPMPointLoadCondition& rOther);
-
-    /// Copy constructor.
-    //MPMPointLoadCondition(const MPMPointLoadCondition& rOther);
-
-
-    ///@}
 
 }; // Class MPMPointLoadCondition
 
@@ -259,25 +243,8 @@ private:
 ///@name Input and output
 ///@{
 
-
-/// input stream function
-/*  inline std::istream& operator >> (std::istream& rIStream,
-        MPMPointLoadCondition& rThis);
-*/
-/// output stream function
-/*  inline std::ostream& operator << (std::ostream& rOStream,
-        const MPMPointLoadCondition& rThis)
-    {
-      rThis.PrintInfo(rOStream);
-      rOStream << std::endl;
-      rThis.PrintData(rOStream);
-
-      return rOStream;
-    }*/
-///@}
-
 }  // namespace Kratos.
 
-#endif // KRATOS_MPM_POINT_LOAD_CONDITION_H_INCLUDED  defined 
+#endif // KRATOS_MPM_GRID_POINT_LOAD_CONDITION_H_INCLUDED  defined
